drawing event dots
